+++ src/visit.rs
@@ -11,9 +11,8 @@ use std::collections::VecDeque;
 use std::sync::Arc;
 
 use anyhow::Context;
-use itertools::Itertools;
-use proc_macro2::Ident;
-use quote::quote;
+use proc_macro2::{Ident, TokenStream};
+use quote::{quote, ToTokens};
 use syn::ext::IdentExt;
 use syn::spanned::Spanned;
 use syn::visit::Visit;
@@ -33,7 +32,7 @@ use crate::*;
 /// were visited but that produced no mutants.
 pub struct Discovered {
     pub mutants: Vec<Mutant>,
-    pub files: Vec<Arc<SourceFile>>,
+    pub files: Vec<SourceFile>,
 }
 
 /// Discover all mutants and all source files.
@@ -42,7 +41,7 @@ pub struct Discovered {
 ///
 pub fn walk_tree(
     workspace_dir: &Utf8Path,
-    top_source_files: &[Arc<SourceFile>],
+    top_source_files: &[SourceFile],
     options: &Options,
     console: &Console,
 ) -> Result<Discovered> {
@@ -52,24 +51,24 @@ pub fn walk_tree(
         .map(|e| syn::parse_str(e).with_context(|| format!("Failed to parse error value {e:?}")))
         .collect::<Result<Vec<Expr>>>()?;
     console.walk_tree_start();
-    let mut file_queue: VecDeque<Arc<SourceFile>> = top_source_files.iter().cloned().collect();
+    let mut file_queue: VecDeque<SourceFile> = top_source_files.iter().cloned().collect();
     let mut mutants = Vec::new();
-    let mut files: Vec<Arc<SourceFile>> = Vec::new();
+    let mut files: Vec<SourceFile> = Vec::new();
     while let Some(source_file) = file_queue.pop_front() {
         console.walk_tree_update(files.len(), mutants.len());
         check_interrupted()?;
-        let (mut file_mutants, external_mods) = walk_file(Arc::clone(&source_file), &error_exprs)?;
+        let (mut file_mutants, external_mods) = walk_file(&source_file, &error_exprs)?;
         // We'll still walk down through files that don't match globs, so that
         // we have a chance to find modules underneath them. However, we won't
         // collect any mutants from them, and they don't count as "seen" for
         // `--list-files`.
         for mod_name in &external_mods {
             if let Some(mod_path) = find_mod_source(workspace_dir, &source_file, mod_name)? {
-                file_queue.push_back(Arc::new(SourceFile::new(
+                file_queue.push_back(SourceFile::new(
                     workspace_dir,
                     mod_path,
                     &source_file.package,
-                )?))
+                )?)
             }
         }
         let path = &source_file.tree_relative_path;
@@ -101,13 +100,10 @@ pub fn walk_tree(
 ///
 /// Returns the mutants found, and the names of modules referenced by `mod` statements
 /// that should be visited later.
-fn walk_file(
-    source_file: Arc<SourceFile>,
-    error_exprs: &[Expr],
-) -> Result<(Vec<Mutant>, Vec<String>)> {
+fn walk_file(source_file: &SourceFile, error_exprs: &[Expr]) -> Result<(Vec<Mutant>, Vec<String>)> {
     let _span = debug_span!("source_file", path = source_file.tree_relative_slashes()).entered();
     debug!("visit source file");
-    let syn_file = syn::parse_str::<syn::File>(&source_file.code)
+    let syn_file = syn::parse_str::<syn::File>(source_file.code())
         .with_context(|| format!("failed to parse {}", source_file.tree_relative_slashes()))?;
     let mut visitor = DiscoveryVisitor {
         error_exprs,
@@ -115,7 +111,7 @@ fn walk_file(
         mutants: Vec::new(),
         namespace_stack: Vec::new(),
         fn_stack: Vec::new(),
-        source_file: Arc::clone(&source_file),
+        source_file: source_file.clone(),
     };
     visitor.visit_file(&syn_file);
     Ok((visitor.mutants, visitor.external_mods))
@@ -131,7 +127,7 @@ struct DiscoveryVisitor<'o> {
     mutants: Vec<Mutant>,
 
     /// The file being visited.
-    source_file: Arc<SourceFile>,
+    source_file: SourceFile,
 
     /// The stack of namespaces we're currently inside.
     namespace_stack: Vec<String>,
@@ -175,26 +171,46 @@ impl<'o> DiscoveryVisitor<'o> {
         );
     }
 
+    /// Record that we generated some mutants.
+    fn collect_mutant(&mut self, span: Span, replacement: TokenStream, genre: Genre) {
+        self.mutants.push(Mutant {
+            source_file: self.source_file.clone(),
+            function: self.fn_stack.last().map(Arc::clone),
+            span,
+            replacement: replacement.to_pretty_string(),
+            genre,
+        })
+    }
+
     fn collect_fn_mutants(&mut self, sig: &Signature, block: &Block) {
-        if let Some(function) = self.fn_stack.last() {
+        if let Some(function) = self.fn_stack.last().map(Arc::clone) {
             let body_span = function_body_span(block).expect("Empty function body");
-            let mut new_mutants = return_type_replacements(&sig.output, self.error_exprs)
-                .map(|rep| Mutant {
-                    source_file: Arc::clone(&self.source_file),
-                    function: Some(Arc::clone(function)),
-                    span: body_span,
-                    replacement: rep.to_pretty_string(),
-                    genre: Genre::FnValue,
-                })
-                .collect_vec();
-            if new_mutants.is_empty() {
+            let repls = return_type_replacements(&sig.output, self.error_exprs);
+            if repls.is_empty() {
                 debug!(
                     function_name = function.function_name,
                     return_type = function.return_type,
                     "No mutants generated for this return type"
                 );
             } else {
-                self.mutants.append(&mut new_mutants);
+                let orig_block = block.to_token_stream().to_pretty_string();
+                for rep in repls {
+                    // Comparing strings is a kludge for proc_macro2 not (yet) apparently
+                    // exposing any way to compare token streams...
+                    //
+                    // TODO: Maybe this should move into collect_mutant, but at the moment
+                    // FnValue is the only genre that seems able to generate no-ops.
+                    //
+                    // The original block has braces and the replacements don't, so put
+                    // them back for the comparison...
+                    let new_block = quote!( { #rep } ).to_token_stream().to_pretty_string();
+                    // dbg!(&orig_block, &new_block);
+                    if orig_block == new_block {
+                        debug!("Replacement is the same as the function body; skipping");
+                    } else {
+                        self.collect_mutant(body_span, rep, Genre::FnValue);
+                    }
+                }
             }
         } else {
             warn!("collect_fn_mutants called while not in a function?");
@@ -350,37 +366,26 @@ impl<'ast> Visit<'ast> for DiscoveryVisitor<'_> {
             BinOp::Gt(_) => vec![quote! { == }, quote! {<}],
             BinOp::Le(_) => vec![quote! {>}],
             BinOp::Ge(_) => vec![quote! {<}],
-            _ => Vec::new(),
+            _ => {
+                trace!(
+                    op = i.op.to_pretty_string(),
+                    "No mutants generated for this binary operator"
+                );
+                Vec::new()
+            }
         };
-        let mut new_mutants = replacements
+        replacements
             .into_iter()
-            .map(|rep| Mutant {
-                source_file: Arc::clone(&self.source_file),
-                function: self.fn_stack.last().map(Arc::clone),
-                replacement: rep.to_pretty_string(),
-                span: i.op.span().into(),
-                genre: Genre::BinaryOperator,
-            })
-            .collect_vec();
-        if new_mutants.is_empty() {
-            debug!(
-                op = i.op.to_pretty_string(),
-                "No mutants generated for this binary operator"
-            );
-        } else {
-            self.mutants.append(&mut new_mutants);
-        }
+            .for_each(|rep| self.collect_mutant(i.op.span().into(), rep, Genre::BinaryOperator));
         syn::visit::visit_expr_binary(self, i);
     }
 }
 
 // Get the span of the block excluding the braces, or None if it is empty.
 fn function_body_span(block: &Block) -> Option<Span> {
-    let start = block.stmts.first()?.span().start();
-    let end = block.stmts.last()?.span().end();
     Some(Span {
-        start: start.into(),
-        end: end.into(),
+        start: block.stmts.first()?.span().start().into(),
+        end: block.stmts.last()?.span().end().into(),
     })
 }
 
@@ -515,7 +520,37 @@ fn attr_is_mutants_skip(attr: &Attribute) -> bool {
 
 #[cfg(test)]
 mod test {
+    use indoc::indoc;
+    use itertools::Itertools;
+
     use super::*;
+    use crate::package::Package;
+    use crate::source::SourceFile;
+
+    /// We should not generate mutants that produce the same tokens as the
+    /// source.
+    #[test]
+    fn no_mutants_equivalent_to_source() {
+        let code = indoc! { "
+            fn always_true() -> bool { true }
+        "};
+        let source_file = SourceFile {
+            code: Arc::new(code.to_owned()),
+            package: Arc::new(Package {
+                name: "unimportant".to_owned(),
+                relative_manifest_path: "Cargo.toml".into(),
+            }),
+            tree_relative_path: Utf8PathBuf::from("src/lib.rs"),
+        };
+        let (mutants, _files) = walk_file(&source_file, &[]).expect("walk_file");
+        let mutant_names = mutants.iter().map(|m| m.name(false, false)).collect_vec();
+        // It would be good to suggest replacing this with 'false', breaking a key behavior,
+        // but bad to replace it with 'true', changing nothing.
+        assert_eq!(
+            mutant_names,
+            ["src/lib.rs: replace always_true -> bool with false"]
+        );
+    }
 
     /// As a generic protection against regressions in discovery, the the mutants
     /// generated from `cargo-mutants` own tree against a checked-in list.